third_party/WebKit/LayoutTests/fast/dom/custom/crash-in-getTypeExtension.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(() => {
  // TODO(a.obzhirov): Use customElements.define when customized built-in elements are supported.
  document.registerElement('x-foo', { prototype: Object.create(HTMLDivElement.prototype) });
  var toStringCalled = false;
  document.createElement('div', { toString: () => { toStringCalled = true; return 'x-foo'; } })
  assert_false(toStringCalled, 'toString should not have been called.');
}, 'Test for crbug/666610: Crash in blink::getTypeExtension');
</script>
